+++ Source/fluffy-obd-lib/diagnostics-engine/elm327/elm327-data-parser.cpp
@@ -12,7 +12,7 @@ RawResponse Elm327DataParser::ParseResponse(std::string data,
         + ")([0-9A-F]+)" };
 
     if (std::regex_search(data, match, dataBytesPattern)) {
-        if (match[4].length() != dataBytesCount * 2) {
+        if (static_cast<size_t>(match[4].length()) != dataBytesCount * 2) {
             throw(std::runtime_error { "Response to: " + commandPid.substr(0, commandPid.size() - 1) + ", not found!\n" });
         }
 
@@ -20,7 +20,7 @@ RawResponse Elm327DataParser::ParseResponse(std::string data,
         parsedResponse.m_commandId = match[3];
         parsedResponse.m_data = match[4];
         parsedResponse.m_ecuId = match[1];
-        parsedResponse.m_length = std::stoi(match[2]);
+        parsedResponse.m_length = static_cast<uint8_t>(std::stoi(match[2]));
 
         return parsedResponse;
     }
